Read job status entirely from the database
Currently we cache job information in memory. However if the ConTest
instance is restarted, or if it's not the owner of a specific job, it
will not be able to reply to a status request.
Reading the database directly will allow any instance with access to the
database to report the status of a job, regardless of the above
conditions. This will enable reliable access to historical data, and
sharding of the ConTest instances. This is also a building block to
enable job resumption.
This introduces a harder dependency on the DB, which should be fair as
the status ownership should always be in the DB, especially valid for
sharded deployments.

Signed-off-by: Andrea Barberio <barberio@fb.com>